git commit -m "feat: add teams + API documentation
- Team, TeamMember, ProjectTeam entities (keep ProjectMember for permissions)
- Issue assignee supports User OR Team
- Updated API docs with request/response examples"

+++ src/main/java/com/rudra/issue_tracker/repository/IssueStatusRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface IssueStatusRepository extends JpaRepository<IssueStatus, Long> {
-
+public interface IssueStatusRepository extends JpaRepository<IssueStatus, Integer> {
     Optional<IssueStatus> findByName(String name);
 }
